+++ backend/Rhyolite/Backend/App.hs
@@ -13,11 +13,7 @@
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE UndecidableInstances #-}
 
-module Rhyolite.Backend.App
-  ( module Rhyolite.Backend.App
-  -- re-export
-  , Postgresql
-  ) where
+module Rhyolite.Backend.App where
 
 import Control.Category (Category)
 import qualified Control.Category as Cat
@@ -41,20 +37,18 @@ import qualified Data.Text.IO as T
 import Data.Typeable (Typeable)
 import Data.Witherable (Filterable(..))
 import Debug.Trace (trace)
-import Database.Groundhog.Postgresql (Postgresql (..))
 import qualified Database.PostgreSQL.Simple as Pg
 import Reflex.Query.Base (mapQuery, mapQueryResult)
 import Reflex.Query.Class (Query, QueryResult, QueryMorphism (..), SelectedCount (..), crop)
 import Snap.Core (MonadSnap, Snap)
 import qualified Web.ClientSession as CS
 import qualified Network.WebSockets as WS
-import Data.Coerce (coerce)
 import Data.Vessel
 import Reflex (Group(..), Additive)
 
 import Rhyolite.Api
 import Rhyolite.App
-import Rhyolite.Backend.Listen (startNotificationListener)
+import Rhyolite.DB.NotifyListen (startNotificationListener, defaultNotificationChannel)
 import Rhyolite.Concurrent
 import Rhyolite.Sign (Signed)
 import Rhyolite.Backend.WebSocket (withWebsocketsConnection, getDataMessage, sendEncodedDataMessage)
@@ -411,7 +405,7 @@ serveDbOverWebsockets
      , Additive q'
      , PositivePart q'
      )
-  => Pool Postgresql
+  => Pool Pg.Connection
   -> RequestHandler r IO
   -> (notifyMessage -> q' -> IO (QueryResult q'))
   -> QueryHandler q' IO
@@ -445,21 +439,18 @@ serveDbOverWebsocketsRaw
   => ((WS.Connection -> IO ()) -> m a)
   -> Text -- ^ version
   -> QueryMorphism qWire q -- ^ Query morphism to translate between wire queries and queries with a reasonable group instance. cf. functorFromWire, vesselFromWire
-  -> Pool Postgresql
+  -> Pool Pg.Connection
   -> RequestHandler r IO
   -> (notifyMessage -> q' -> IO (QueryResult q'))
   -> QueryHandler q' IO
   -> Pipeline IO (MonoidalMap ClientKey q) q'
   -> IO (m a, IO ())
 serveDbOverWebsocketsRaw withWsConn version fromWire db handleApi handleNotify handleQuery pipe = do
-  (getNextNotification, finalizeListener) <- startNotificationListener db
+  (getNextNotification, finalizeListener) <- startNotificationListener defaultNotificationChannel db
   rec (qh, finalizeFeed) <- feedPipeline (handleNotify <$> getNextNotification) handleQuery r
       (qh', r) <- unPipeline pipe qh r'
       (r', handleListen) <- connectPipelineToWebsocketsRaw withWsConn version fromWire handleApi qh'
   return (handleListen, finalizeFeed >> finalizeListener)
-
-convertPostgresPool :: Pool Pg.Connection -> Pool Postgresql
-convertPostgresPool = coerce
 
 -- | This is typically useful to provide as a last argument to serveDbOverWebsockets, as it handles
 -- the combinatorics of aggregating the queries of connected clients as provided to the handler for